uudeview 0.5.20
UUDeview is a program that helps you transmit and receive binary files
over the Internet, using electronic mail or newsgroups. The UUDeview
package includes both an encoder and a decoder. The decoder
automatically detects the type of encoding used, offering MIME's Base64
and BinHex as well as the popular uuencoding and the less frequently
used xxencoding methods.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>